Commands folder structure updated
