+++ src/main/java/frc/robot/commands/arm/Cmd_ArmCargoMid.java
@@ -5,40 +5,43 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_ArmHome extends Command {
-  public Cmd_ArmHome() {
-    // Use requires() here to declare subsystem dependencies
+
+public class Cmd_ArmCargoMid extends Command {
+  float sign;
+  int distance = 0;
+
+
+  public Cmd_ArmCargoMid() {
     requires(Robot.s_arm);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
+    setTimeout(5);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_arm.dartDriveGoDown();
+    Robot.s_arm.setArmPosTest(Robot.s_arm.midCargoPos);
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return(Robot.s_arm.isArmHome());
+    return(Robot.s_arm.dartMotor0Position() == Robot.s_arm.ARM_TICKS * Robot.s_arm.midCargoPos || isTimedOut());
   }
 
   // Called once after isFinished returns true
   @Override
   protected void end() {
     Robot.s_arm.armStop();
-    Robot.s_arm.resetArmPos();
-
   }
 
   // Called when another command which requires one or more of the same
@@ -46,6 +49,5 @@ public class Cmd_ArmHome extends Command {
   @Override
   protected void interrupted() {
     Robot.s_arm.armStop();
-
   }
 }
