+++ src/main/java/frc/robot/commands/arm/Cmd_ArmHatchHigh.java
@@ -5,44 +5,49 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_ManualBars extends Command {
 
-  public Cmd_ManualBars() {
-    requires(Robot.s_bars);
+public class Cmd_ArmHatchHigh extends Command {
+  float sign;
+  int distance = 0;
+
+
+  public Cmd_ArmHatchHigh() {
+    requires(Robot.s_arm);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
+    setTimeout(5);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_bars.barsManualSpeed(0.5);
+    Robot.s_arm.setArmPosTest(Robot.s_arm.highHatchPos);
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return false;
+    return(Robot.s_arm.dartMotor0Position() == Robot.s_arm.ARM_TICKS * Robot.s_arm.highHatchPos || isTimedOut());
   }
 
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_bars.barsManualSpeed(0);
+    Robot.s_arm.armStop();
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_bars.barsManualSpeed(0);
+    Robot.s_arm.armStop();
   }
 }
