+++ src/main/java/frc/robot/commands/arm/Cmd_ArmHatchMid.java
@@ -5,45 +5,49 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_CollectorForwardExtend extends Command {
-  public Cmd_CollectorForwardExtend() {
-    requires(Robot.s_collector);
-    // Use requires() here to declare subsystem dependencies
-    // eg. requires(chassis);
+
+public class Cmd_ArmHatchMid extends Command {
+  float sign;
+  int distance = 0;
+
+
+  public Cmd_ArmHatchMid() {
+    requires(Robot.s_arm);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
+    setTimeout(5);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_collector.solenoidExtendCollector();
+    Robot.s_arm.setArmPosTest(Robot.s_arm.midHatchPos);
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return (Robot.s_collector.isBallSensor());
+    return(Robot.s_arm.dartMotor0Position() == Robot.s_arm.ARM_TICKS * Robot.s_arm.midHatchPos || isTimedOut());
   }
 
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_collector.solenoidRetractCollector();
+    Robot.s_arm.armStop();
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_collector.solenoidRetractCollector();
+    Robot.s_arm.armStop();
   }
 }
