+++ src/main/java/frc/robot/commands/arm/Cmd_ArmManualDown.java
@@ -5,37 +5,31 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-
-public class Cmd_ArmCargoShip extends Command {
-  float sign;
-  int distance = 0;
-
-
-  public Cmd_ArmCargoShip() {
-    requires(Robot.s_arm);
+public class Cmd_ArmManualDown extends Command {
+  public Cmd_ArmManualDown() {
+  requires(Robot.s_arm);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
-    setTimeout(5);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_arm.setArmPosTest(Robot.s_arm.shipCargoPos);
+    Robot.s_arm.dartDriveGoDown();
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return(Robot.s_arm.dartMotor0Position() == Robot.s_arm.ARM_TICKS * Robot.s_arm.shipCargoPos || isTimedOut());
+    return false;
   }
 
   // Called once after isFinished returns true
