+++ src/main/java/frc/robot/commands/arm/Cmd_ArmManualUp.java
@@ -5,37 +5,31 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.arm;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-
-public class Cmd_ArmHatchHigh extends Command {
-  float sign;
-  int distance = 0;
-
-
-  public Cmd_ArmHatchHigh() {
+public class Cmd_ArmManualUp extends Command {
+  public Cmd_ArmManualUp() {
     requires(Robot.s_arm);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
-    setTimeout(5);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_arm.setArmPosTest(Robot.s_arm.highHatchPos);
+    Robot.s_arm.dartDriveGoUp();
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return(Robot.s_arm.dartMotor0Position() == Robot.s_arm.ARM_TICKS * Robot.s_arm.highHatchPos || isTimedOut());
+    return false;
   }
 
   // Called once after isFinished returns true
@@ -49,5 +43,6 @@ public class Cmd_ArmHatchHigh extends Command {
   @Override
   protected void interrupted() {
     Robot.s_arm.armStop();
+
   }
 }
