+++ src/main/java/frc/robot/commands/collector/Cmd_CollectorDropMotors.java
@@ -5,16 +5,17 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.collector;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_CollectorInput extends Command {
-  public Cmd_CollectorInput() {
+public class Cmd_CollectorDropMotors extends Command {
+  double speed;
+
+public Cmd_CollectorDropMotors(double speed) {
     requires(Robot.s_collector);
-    // Use requires() here to declare subsystem dependencies
-    // eg. requires(chassis);
+    this.speed = speed;
   }
 
   // Called just before this Command runs the first time
@@ -25,25 +26,27 @@ public class Cmd_CollectorInput extends Command {
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_collector.collectorIntake();
+    Robot.s_collector.dropMotorSetSpeed(speed);
+
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
     return (Robot.s_collector.isBallSensor());
-  }
-
+}
+  
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_collector.collectorStop();
+    Robot.s_collector.dropMotorSetSpeed(0);
+
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_collector.collectorStop();
+    Robot.s_collector.dropMotorSetSpeed(0);
   }
 }
