+++ src/main/java/frc/robot/commands/collector/Cmd_CollectorForwardExtend.java
@@ -5,14 +5,16 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.collector;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_TestDropRun extends Command {
-  public Cmd_TestDropRun() {
+public class Cmd_CollectorForwardExtend extends Command {
+  public Cmd_CollectorForwardExtend() {
     requires(Robot.s_collector);
+    // Use requires() here to declare subsystem dependencies
+    // eg. requires(chassis);
   }
 
   // Called just before this Command runs the first time
@@ -23,27 +25,25 @@ public class Cmd_TestDropRun extends Command {
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_collector.dropMotorSetSpeed(.5);
+    Robot.s_collector.solenoidExtendCollector();
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return false;
+    return (Robot.s_collector.isBallSensor());
   }
 
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_collector.dropMotorSetSpeed(0);
-
+    Robot.s_collector.solenoidRetractCollector();
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_collector.dropMotorSetSpeed(0);
-
+    Robot.s_collector.solenoidRetractCollector();
   }
 }
