+++ src/main/java/frc/robot/commands/collector/Cmd_CollectorInput.java
@@ -5,67 +5,45 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.collector;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_AutoDrive extends Command {
-
-  int distance = 0;
-  double upperSpeed;
-  double lowerSpeed;
-  float sign;
-
-  public Cmd_AutoDrive(int distance, double upperSpeed, double lowerSpeed) {
-    requires(Robot.s_drivetrain);
-    this.distance = distance;
-    this.upperSpeed = upperSpeed;
-    this.lowerSpeed = lowerSpeed;
+public class Cmd_CollectorInput extends Command {
+  public Cmd_CollectorInput() {
+    requires(Robot.s_collector);
+    // Use requires() here to declare subsystem dependencies
+    // eg. requires(chassis);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
-    Robot.s_drivetrain.resetGyro();
-    Robot.s_drivetrain.setSetpointPos(distance);
-    setTimeout(15);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_drivetrain.driveToPos(upperSpeed, lowerSpeed);
+    Robot.s_collector.collectorIntake();
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    sign = Math.signum(distance);
-        
-        if (sign == 1){
-            return Robot.s_drivetrain.isDoneDriving() || isTimedOut();
-        }
-        if (sign == -1){
-            return Robot.s_drivetrain.isDoneDrivingBack() || isTimedOut();
-        }
-        else{
-            return false;
-        }
+    return (Robot.s_collector.isBallSensor());
   }
 
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_drivetrain.driveStop();
-
+    Robot.s_collector.collectorStop();
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_drivetrain.driveStop();
-
+    Robot.s_collector.collectorStop();
   }
 }
