+++ src/main/java/frc/robot/commands/collector/Cmd_CollectorThrow.java
@@ -5,49 +5,46 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.collector;
 
-import edu.wpi.first.wpilibj.command.Command;
+import edu.wpi.first.wpilibj.command.TimedCommand;
 import frc.robot.Robot;
 
-
-public class Cmd_ArmCargoLow extends Command {
-  float sign;
-  int distance = 0;
-
-
-  public Cmd_ArmCargoLow() {
-    requires(Robot.s_arm);
+/**
+ * Add your docs here.
+ */
+public class Cmd_CollectorThrow extends TimedCommand {
+  /**
+   * Add your docs here.
+   */
+  public Cmd_CollectorThrow(double timeout) {
+    super(timeout);
+      requires(Robot.s_collector);
+    // Use requires() here to declare subsystem dependencies
+    // eg. requires(chassis);
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
-    setTimeout(5);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_arm.setArmPosTest(Robot.s_arm.lowCargoPos);
+    Robot.s_collector.collectorThrow();
   }
 
-  // Make this return true when this Command no longer needs to run execute()
-  @Override
-  protected boolean isFinished() {
-    return(Robot.s_arm.dartMotor0Position() == Robot.s_arm.ARM_TICKS * Robot.s_arm.lowCargoPos || isTimedOut());
-  }
-
-  // Called once after isFinished returns true
+  // Called once after timeout
   @Override
   protected void end() {
-    Robot.s_arm.armStop();
+    Robot.s_collector.collectorStop();
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_arm.armStop();
+    Robot.s_collector.collectorStop();
   }
 }
