+++ src/main/java/frc/robot/commands/drive/Cmd_AutoDrive.java
@@ -5,41 +5,59 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.drive;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_CollectorDropMotors extends Command {
-  double speed;
+public class Cmd_AutoDrive extends Command {
 
-public Cmd_CollectorDropMotors(double speed) {
-    requires(Robot.s_collector);
-    this.speed = speed;
+  int distance = 0;
+  double upperSpeed;
+  double lowerSpeed;
+  float sign;
+
+  public Cmd_AutoDrive(int distance, double upperSpeed, double lowerSpeed) {
+    requires(Robot.s_drivetrain);
+    this.distance = distance;
+    this.upperSpeed = upperSpeed;
+    this.lowerSpeed = lowerSpeed;
   }
 
   // Called just before this Command runs the first time
   @Override
   protected void initialize() {
+    Robot.s_drivetrain.resetGyro();
+    Robot.s_drivetrain.setSetpointPos(distance);
+    setTimeout(15);
   }
 
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_collector.dropMotorSetSpeed(speed);
-
+    Robot.s_drivetrain.driveToPos(upperSpeed, lowerSpeed);
   }
 
   // Make this return true when this Command no longer needs to run execute()
   @Override
   protected boolean isFinished() {
-    return (Robot.s_collector.isBallSensor());
-}
-  
+    sign = Math.signum(distance);
+        
+        if (sign == 1){
+            return Robot.s_drivetrain.isDoneDriving() || isTimedOut();
+        }
+        if (sign == -1){
+            return Robot.s_drivetrain.isDoneDrivingBack() || isTimedOut();
+        }
+        else{
+            return false;
+        }
+  }
+
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_collector.dropMotorSetSpeed(0);
+    Robot.s_drivetrain.driveStop();
 
   }
 
@@ -47,6 +65,7 @@ public Cmd_CollectorDropMotors(double speed) {
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_collector.dropMotorSetSpeed(0);
+    Robot.s_drivetrain.driveStop();
+
   }
 }
