+++ src/main/java/frc/robot/commands/drive/Cmd_ManualDrive.java
@@ -5,14 +5,20 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.drive;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.Robot;
 
-public class Cmd_ArmManualDown extends Command {
-  public Cmd_ArmManualDown() {
-  requires(Robot.s_arm);
+public class Cmd_ManualDrive extends Command {
+
+  private double xCube;
+  private double Kp = -0.05;
+  private double min_command = 0.3;
+
+  public Cmd_ManualDrive() {
+    requires(Robot.s_drivetrain);
+    requires(Robot.s_limelight);
   }
 
   // Called just before this Command runs the first time
@@ -23,7 +29,29 @@ public class Cmd_ArmManualDown extends Command {
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    Robot.s_arm.dartDriveGoDown();
+    if (Robot.m_oi.getButtonState(7) && Robot.m_oi.getButtonState(8)) {
+      xCube = Robot.s_limelight.getCubeX();
+
+      double heading_error = -xCube;
+      double steering_adjust = 0.0;
+
+          if(xCube > 1){
+              steering_adjust = Kp * heading_error + min_command;
+          }
+          if(xCube < 1){
+              steering_adjust = Kp * heading_error - min_command;
+          }
+
+      Robot.s_drivetrain.trackCubeManualSpeed(steering_adjust, -Robot.m_oi.getGamepad().getY());
+      }
+
+      else if (Robot.m_oi.getButtonState(8)){
+      Robot.s_drivetrain.driveArcade(Robot.m_oi.getGamepad());
+      }
+
+          else {
+              Robot.s_drivetrain.drive(Robot.m_oi.getGamepad());
+          }
   }
 
   // Make this return true when this Command no longer needs to run execute()
@@ -35,13 +63,11 @@ public class Cmd_ArmManualDown extends Command {
   // Called once after isFinished returns true
   @Override
   protected void end() {
-    Robot.s_arm.armStop();
   }
 
   // Called when another command which requires one or more of the same
   // subsystems is scheduled to run
   @Override
   protected void interrupted() {
-    Robot.s_arm.armStop();
   }
 }
